refactor: convert more stories to TS
also useGLTF does not need to declare it's type
